requirements.txt
Babel==2.3.4
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-activity-stream==0.6.3
django-angular==0.8.4
django-countries==4.0
django-crispy-forms==1.6.0
django-easy-pdf==0.1.0
django-excel==0.0.6
django-google-charts==0.1.1
django-jsonfield==1.0.1
django-phonenumber-field==1.1.0
django-qsstats-magic==0.7.2
djangorestframework==3.4.0
et-xmlfile==1.0.1
future==0.16.0
gevent==1.1.1
greenlet==0.4.10
gunicorn==19.6.0
html5lib==0.9999999
jdcal==1.3
openpyxl==2.4.0
phonenumberslite==7.7.4
Pillow==3.2.0
psycopg2==2.6.1
pyexcel==0.3.3
pyexcel-io==0.2.3
pyexcel-webio==0.0.8
pyexcel-xls==0.2.3
pyexcel-xlsx==0.2.3
PyPDF2==1.26.0
python-dateutil==2.5.3
pytz==2016.7
reportlab==3.3.0
six==1.10.0
static3==0.7.0
texttable==0.8.7
virtualenv==15.0.2
whitenoise==3.2
xhtml2pdf==0.0.6
xlrd==1.0.0
xlwt-future==0.8.0
